refactor : FrameInfo assembleFrameInfo 메서드 구현
